pyproject.toml
[project]
name = "stock-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
